adding GPU 2022 benchmark raster seeding

+++ Models/DTOs/CPU_Benchmark.cs
@@ -8,14 +8,13 @@ namespace pcp2p.Models
     public class CPU_Benchmark
     {
         public string Name {get;set;}
-        public int FPS {get;set;}        
+        public float FPS {get;set;}        
     }
 
     public sealed class CPU_Benchmark_Map : CsvClassMap<CPU_Benchmark>
     {
         public CPU_Benchmark_Map()
         {
-            // We ignore "1024" and "FPS" by simply not mapping them
             Map(m => m.Name).Name("CPU Name");
             Map(m => m.FPS).Name("FPS");
         }
